+++ client/src/ui.rs
@@ -9,38 +9,7 @@ use shared::units::Unit;
 
 use crate::LocalPlayerColor;
 use crate::input::{Controller, InputSelection, LastSubmittedTurn, TargetableVerb, UiState};
-
-pub struct ColorState {
-    pub idle: Color,
-    pub hover: Color,
-    pub pressed: Color,
-    pub waiting: Color,
-}
-
-pub struct ButtonTheme {
-    pub background: ColorState,
-    pub border: ColorState,
-}
-
-pub mod theme {
-    use super::*;
-    use bevy::color::palettes::css::*;
-
-    pub const FINISH_BUTTON: ButtonTheme = ButtonTheme {
-        background: ColorState {
-            idle: Color::Srgba(DARK_CYAN),
-            hover: Color::Srgba(LIGHT_CYAN),
-            pressed: Color::Srgba(DARK_SLATE_GRAY),
-            waiting: Color::Srgba(DARK_GRAY),
-        },
-        border: ColorState {
-            idle: Color::Srgba(TEAL),
-            hover: Color::Srgba(AQUAMARINE),
-            pressed: Color::Srgba(SLATE_GRAY),
-            waiting: Color::Srgba(GRAY),
-        },
-    };
-}
+use crate::visuals::theme;
 
 #[derive(Component)]
 pub struct TurnUiText;
